ex5/requirements.txt
Flask
Flask-JWT-Extended
Flask-RESTful
Flask-SQLAlchemy
Flask-Marshmallow
marshmallow
marshmallow-sqlalchemy